add binIon routine to similarity module for flexibility
Former-commit-id: cd709de2a25fa27450fc5db1cf6c479e0b854f83

+++ backend/services/similarity/src/main/scala/edu/ucdavis/fiehnlab/mona/core/similarity/util/SpectrumUtils.scala
@@ -28,7 +28,7 @@ object SpectrumUtils {
     * @param mz
     * @return
     */
-  def roundMZ(mz: Double): Long = scala.math.round(mz + 0.2 + EPS_CORRECTION)
+  def roundMZ(mz: Double): Int = scala.math.round(mz + 0.2 + EPS_CORRECTION).toInt
 
 
   /**
